Added 'I' command to control the input.  The selection should be for a directory which contains the MOVES and TMP directories, in other words, the top level directory of a design, or the sub-directory in the case of a multi-part design.
When the app is started from the command line (how and where is
inconvenient at this point), the command line argument is the
directory.

The title on the window reflects the file being displayed.

Perhaps TODO:  add a way to display multiple files, without
needing to merge them ahead of time.